Refactor game initilization
Create a static MapCatalog class that hold different fort configurations
This will eventually be turned into a json deserializer
that will read fort configurations in from json text files.

The Game class holds the logic to run a game. Determining who the
winner is and telling the bots to make their moves.

The Map class contains a list of forts and some logic to find out
the players in those forts.

Eventually there will be a tournament class that will take in a json
file that describes which bots and which maps should be used. It will
then generate and run the needed game objects.